requirements.txt
python-telegram-bot>=20.0
openai>=1.0.0
httpx>=0.24.0
pillow